Add BigDecimal, optimize database requests

+++ app/src/main/java/com/juliandobrodolac/honorarcraftandroid/InvoiceEntry.kt
@@ -2,12 +2,13 @@ package com.juliandobrodolac.honorarcraftandroid
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity(tableName = "invoice_entries")
 data class InvoiceEntry(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     var invoiceNumber: String, // Foreign key to link to an InvoiceData
     val date: String,
-    val lessonUnits: Double,
+    val lessonUnits: BigDecimal,
     val teachingSubject: String
 )
